requirements.txt

streamlit==1.38.0
beautifulsoup4==4.12.3
requests==2.32.3
numpy==2.3.3
scipy==1.16.3
scikit-learn==1.7.2
pdfminer.six==20231228
mammoth==1.6.0
